Move projects to its own app
The model is still in builds because it needs db changes that need
planning.
