+++ frigg/builds/api.py
@@ -1,13 +1,11 @@
 # -*- coding: utf8 -*-
 import json
 
-from django.http import HttpResponse
 from django.http.response import JsonResponse
 from django.shortcuts import get_object_or_404, render
-from django.views.decorators.cache import never_cache
 from django.views.decorators.csrf import csrf_exempt
 
-from .models import Build, Project
+from .models import Build
 
 
 @csrf_exempt
@@ -21,22 +19,6 @@ def report_build(request):
         response = JsonResponse({'error': 'Build not found'})
         response.status_code = 404
     return response
-
-
-@never_cache
-@csrf_exempt
-def build_badge(request, owner, name, branch='master'):
-    project = get_object_or_404(Project, owner=owner, name=name)
-    badge = project.get_badge(branch)
-    return HttpResponse(content=badge, content_type='image/svg+xml')
-
-
-@never_cache
-@csrf_exempt
-def coverage_badge(request, owner, name, branch='master'):
-    project = get_object_or_404(Project, owner=owner, name=name)
-    badge = project.get_coverage_badge(branch)
-    return HttpResponse(content=badge, content_type='image/svg+xml')
 
 
 def partial_build_page(request, owner, name, build_number):
