+++ frigg/builds/filters.py
@@ -1,11 +1,6 @@
 from rest_framework import filters
 
-from frigg.builds.models import Build, Project
-
-
-class ProjectPermissionFilter(filters.BaseFilterBackend):
-    def filter_queryset(self, request, queryset, view):
-        return queryset.filter(Project.objects.permitted_query(request.user)).distinct()
+from frigg.builds.models import Build
 
 
 class BuildPermissionFilter(filters.BaseFilterBackend):
