+++ frigg/builds/views.py
@@ -34,63 +34,6 @@ def overview(request, page=1):
     })
 
 
-def approve_projects(request, project_id=None):
-    if not request.user.is_superuser:
-        raise Http404
-
-    if project_id and request.method == 'POST' and request.POST.get('approve') == "yes":
-        project = Project.objects.get(id=project_id)
-        project.approved = True
-        project.save()
-        if project.builds.all():
-            project.builds.last().start()
-
-        return redirect('approve_projects_overview')
-
-    return render(request, "builds/approve_projects.html", {
-        'projects': Project.objects.filter(approved=False)
-    })
-
-
-def view_organization(request, owner):
-    builds = Build.objects.permitted(request.user) \
-        .filter(project__owner=owner) \
-        .select_related('project', 'result__still_running', 'result__succeeded')
-
-    if len(builds) == 0:
-        raise Http404
-
-    success_rate = 0
-    if len(builds):
-        success_rate = int((builds.filter(result__succeeded=True).count()/len(builds)) * 100)
-
-    return render(request, "builds/organization.html", {
-        'organization': owner,
-        'builds': builds,
-        'success_rate': success_rate
-    })
-
-
-def view_project(request, owner, name):
-    project = get_object_or_404(
-        Project.objects.permitted(request.user).prefetch_related('builds', 'builds__result'),
-        owner=owner,
-        name=name
-    )
-    builds = project.builds.all()\
-                           .select_related('project', 'result__still_running', 'result__succeeded')
-
-    success_rate = 0
-    if len(builds):
-        success_rate = int((builds.filter(result__succeeded=True).count()/len(builds)) * 100)
-
-    return render(request, "builds/project.html", {
-        'project': project,
-        'builds': builds,
-        'success_rate': success_rate
-    })
-
-
 def last_build(request, owner, name):
     project = get_object_or_404(Project.objects.permitted(request.user), owner=owner, name=name)
     return redirect(project.builds.first())
