+++ frigg/projects/filters.py
@@ -1,0 +1,8 @@
+from rest_framework import filters
+
+from .models import Project
+
+
+class ProjectPermissionFilter(filters.BaseFilterBackend):
+    def filter_queryset(self, request, queryset, view):
+        return queryset.filter(Project.objects.permitted_query(request.user)).distinct()
