+++ frigg/projects/models.py
@@ -1,0 +1,1 @@
+from frigg.builds.models import Project  # noqa
